Generalize enso_cloud log API access
