+++ std-bits/base/src/main/java/org/enso/base/enso_cloud/audit/AuditLog.java
@@ -13,16 +13,27 @@ import java.util.concurrent.Future;
  * the meantime, all waiting messages (up to some limit) will be sent in a single request.
  */
 public final class AuditLog {
+  private AuditLog() {}
+
+  private static AuditLogApiAccess apiAccess;
+
+  private static AuditLogApiAccess apiAccess() {
+    if (apiAccess == null) {
+      apiAccess = new AuditLogApiAccess();
+    }
+    return apiAccess;
+  }
+
   /** Schedules the log message to be sent in the next batch, and returns immediately. */
   public static void logAsync(String type, String message, ObjectNode metadata) {
     var event = new AuditLogMessage(type, message, metadata);
-    AuditLogApiAccess.INSTANCE.logWithoutConfirmation(event);
+    apiAccess().logWithoutConfirmation(event);
   }
 
   /** Schedules the log message to be sent in the next batch, and waits until it has been sent. */
   public static void logSynchronously(String type, String message, ObjectNode metadata) {
     var event = new AuditLogMessage(type, message, metadata);
-    Future<Void> future = AuditLogApiAccess.INSTANCE.logWithConfirmation(event);
+    Future<Void> future = apiAccess().logWithConfirmation(event);
     try {
       future.get();
     } catch (ExecutionException | InterruptedException e) {
@@ -37,6 +48,6 @@ public final class AuditLog {
   }
 
   public static void resetCache() {
-    AuditLogApiAccess.INSTANCE.resetCache();
+    apiAccess().resetCache();
   }
 }
