+++ std-bits/base/src/main/java/org/enso/base/enso_cloud/logging/LogMessage.java
@@ -1,0 +1,8 @@
+package org.enso.base.enso_cloud.logging;
+
+public interface LogMessage {
+  /**
+   * @return JSON string representation of the log message.
+   */
+  String payload();
+}
