+++ std-bits/base/src/main/java/org/enso/base/enso_cloud/logging/RequestFailureException.java
@@ -1,0 +1,7 @@
+package org.enso.base.enso_cloud.logging;
+
+public final class RequestFailureException extends RuntimeException {
+  public RequestFailureException(String message, Throwable cause) {
+    super(message, cause);
+  }
+}
